requirements.txt
Django>2
djangorestframework
django-extensions

feedparser
requests
